Implements join/create/close game functionalities
Adds handler classes to create, close and join games.
The create game handler creates a new game if the game doesn't already exist and if the server hasn't reached the max amount of games.
The close game handler closes and removes a game.
The join game handler allows players to join games and select a team.

Also adds a retry mechanism when a game is already being created or closed.